requirements.txt
langchain
langchain-community
pymupdf
google-generativeai
python-dotenv
tiktoken
faiss-cpu
sentence-transformers
fastapi
python-multipart
langchain-huggingface
uvicorn